+++ src/app.py
@@ -32,6 +32,7 @@ from devices import (
     eDiluterWidget,
     AFMWidget,
     TSIWidget,
+    AFCWidget,
     ExampleDeviceWidget
 )