+++ src/plotting/device_plot_configs.py
@@ -848,7 +848,7 @@ class AFMPlotConfig(BasePlotConfig):
         """Store all AFM values."""
         afm_data = self.device.current_data
         plot_data[str(dev_id)+':f'][time_counter] = afm_data.flow
-        plot_data[str(dev_id)+':sf'][time_counter] = afm_data.saturator_flow
+        plot_data[str(dev_id)+':sf'][time_counter] = afm_data.standard_flow
         plot_data[str(dev_id)+':rh'][time_counter] = afm_data.humidity
         plot_data[str(dev_id)+':t'][time_counter] = afm_data.temperature
         plot_data[str(dev_id)+':p'][time_counter] = afm_data.pressure
@@ -910,6 +910,15 @@ class AFMPlotConfig(BasePlotConfig):
                 **config
             }
         return None
+
+
+class AFCPlotConfig(BasePlotConfig):
+    """Plot configuration for AFC (Airmodus Flow Controller)."""
+
+    def get_plot_values(self, dev_id, time_counter, plot_data, data_holder=None):
+        """Store AFC flow value."""
+        afc_data = self.device.current_data
+        plot_data[str(dev_id)][time_counter] = afc_data.flow
 
 
 class ExampleDevicePlotConfig(BasePlotConfig):