+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Airmodus MultiLogger - Software change log
 
+### 0.11.4 - 2026.01.07
+- AFC flow value in status bar
+- AFC error flag setting
+- renamed AFC "Set" tab to "Control"
+- removed AFCSettings, moved flow_setpoint to AFCData
+
+### 0.11.3 - 2025.12.15
+- AFC device type detection
+- AFM value saturator_flow renamed to standard_flow
+
+### 0.11.2 - 2025.12.12
+- changed DTR disable to RTS disable (esp32 reboot problem)
+- set main plot value when adding device (multi-value devices)
+- added send_read_commands() function to AFCWidget
+
+### 0.11.1 - 2025.12.05
+- AFC device type (WIP)
+
+### 0.11.0 - 2025.11.10
+- Refactor merge
+
 ### 0.10.9 - 2025.10.09
 - Removed PSM CPC inlet flow reset (when CPC not connected)
 - Added pulse quality indicator to CPC status tab